Update router.py
Sometimes we create migrations for local development databases but someone adds an invalid import, in this situation the real error is hidden from the developer as only `Can't import models module: True` is returned instead of the real issue: `No module named 'peewee_extra_fields'`. This PR is meant to help prevent others from having to download this library and place it in the editable mode with a debugger in it.. just to find out the library works as expected.